Remove constructor super calls
This is a performance optimization for these utility objects that are
created to represent data, but don't rely on constructor side-effects.

Given the huge number of invocations (32000 every time we deselect a
helix) it is extremely important that the constructors are either very
fast or are avoided completely with Object.create.